core/requirements.txt
autopep8==2.0.1
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
cssselect==1.2.0
cssutils==2.6.0
flake8==6.0.0
idna==3.4
lxml==4.9.2
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
premailer==3.10.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-decouple==3.6
requests==2.28.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
yagmail==0.15.293
